Simulation verbessert

+++ RscpExampleMain.cpp
@@ -719,7 +719,7 @@ typedef struct {
 
 std::array<int, 19> oekofen{2,11,12,13,21,22,23,31,32,33,41,42,43,60,61,73,78,101,102};
 static int x1 = 0;
-static int temp[oekofen.size()];
+static int16_t temp[oekofen.size()];
 static uint8_t tn = 1;
 
 int iModbusTCP_Set(int reg,int val,int tac)
@@ -1738,7 +1738,7 @@ bDischarge = false;
     if (strcmp(e3dc_config.heizung_ip,"0.0.0.0")!=0)
     {
         printf("%c[K\n", 27 );
-        printf("oekofen AT %i HK1 %i %i %i %i %i %i ",temp[0],temp[1],temp[2],temp[3],temp[4],temp[5],temp[6]);
+        printf("oekofen AT %i HK1 %i %i %i %i %i %i ",int(temp[0]),temp[1],temp[2],temp[3],temp[4],temp[5],temp[6]);
         printf("HK2 %i %i %i %i %i %i"
                ,temp[7],temp[8],temp[9],temp[10],temp[11],temp[12]);
         printf("%c[K\n", 27 );
